requirements.txt
ultralytics>=8.3.0
torch
torchvision
pandas
numpy
matplotlib
pyyaml
codecarbon
opencv-python
Pillow
